Move `IsrEnviroment` class definition
